Fix #2408 Refactor commands out of Main.hs
