+++ package.yaml
@@ -197,7 +197,10 @@ library:
   - Stack.Coverage
   - Stack.DefaultColorWhen
   - Stack.Docker
+  - Stack.DockerCmd
   - Stack.Dot
+  - Stack.Eval
+  - Stack.Exec
   - Stack.FileWatch
   - Stack.GhcPkg
   - Stack.Ghci
@@ -218,6 +221,7 @@ library:
   - Stack.Options.Completion
   - Stack.Options.DockerParser
   - Stack.Options.DotParser
+  - Stack.Options.EvalParser
   - Stack.Options.ExecParser
   - Stack.Options.GhcBuildParser
   - Stack.Options.GhciParser
@@ -225,20 +229,26 @@ library:
   - Stack.Options.GlobalParser
   - Stack.Options.HaddockParser
   - Stack.Options.HpcReportParser
+  - Stack.Options.InitParser
   - Stack.Options.LogLevelParser
+  - Stack.Options.LsParser
   - Stack.Options.NewParser
   - Stack.Options.NixParser
   - Stack.Options.PackageParser
+  - Stack.Options.PathParser
   - Stack.Options.ResolverParser
-  - Stack.Options.ScriptParser
   - Stack.Options.SDistParser
+  - Stack.Options.ScriptParser
+  - Stack.Options.SetupParser
   - Stack.Options.TestParser
+  - Stack.Options.UpgradeParser
   - Stack.Options.UploadParser
   - Stack.Options.Utils
   - Stack.Package
   - Stack.PackageDump
   - Stack.Path
   - Stack.Prelude
+  - Stack.Query
   - Stack.Runners
   - Stack.Script
   - Stack.SDist
@@ -249,6 +259,7 @@ library:
   - Stack.Storage.Project
   - Stack.Storage.User
   - Stack.Storage.Util
+  - Stack.Templates
   - Stack.Types.Build
   - Stack.Types.CompilerBuild
   - Stack.Types.Compiler
@@ -264,7 +275,9 @@ library:
   - Stack.Types.SourceMap
   - Stack.Types.TemplateName
   - Stack.Types.Version
+  - Stack.Uninstall
   - Stack.Unpack
+  - Stack.Update
   - Stack.Upgrade
   - Stack.Upload
   - System.Info.ShortPathName
