+++ src/Stack/Build.hs
@@ -5,54 +5,54 @@
 -- | Build the project.
 
 module Stack.Build
-  ( build
+  ( buildCmd
+  , build
   , buildLocalTargets
   , loadPackage
   , mkBaseConfigOpts
-  , queryBuildInfo
   , splitObjsWarning
   ) where
 
-import           Data.Aeson ( Value (Object, Array), (.=), object )
-import qualified Data.Aeson.Key as Key
-import qualified Data.Aeson.KeyMap as KeyMap
-import           Data.List ( (\\), isPrefixOf )
+import           Data.Attoparsec.Args ( EscapingMode (Escaping), parseArgs )
+import           Data.List ( (\\) )
 import           Data.List.Extra ( groupSort )
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map as Map
 import qualified Data.Set as Set
 import qualified Data.Text as T
-import           Data.Text.Encoding ( decodeUtf8 )
-import qualified Data.Text.IO as TIO
-import           Data.Text.Read ( decimal )
-import qualified Data.Vector as V
-import qualified Data.Yaml as Yaml
 import qualified Distribution.PackageDescription as C
 import           Distribution.Types.Dependency ( Dependency (..), depLibraries )
 import           Distribution.Version ( mkVersion )
-import           Path ( parent )
 import           Stack.Build.ConstructPlan ( constructPlan )
 import           Stack.Build.Execute ( executePlan, preFetch, printPlan )
 import           Stack.Build.Installed ( getInstalled, toInstallMap )
 import           Stack.Build.Source ( localDependencies, projectLocalPackages )
+import           Stack.Build.Target ( NeedTargets (..) )
+import           Stack.FileWatch ( fileWatch, fileWatchPoll )
 import           Stack.Package ( resolvePackage )
 import           Stack.Prelude hiding ( loadPackage )
+import           Stack.Runners ( ShouldReexec (..), withConfig, withEnvConfig )
 import           Stack.Setup ( withNewLocalBuildTargets )
 import           Stack.Types.Build
                    ( BaseConfigOpts (..), BuildException (..)
                    , BuildPrettyException (..), Plan (..), Task (..)
                    , TaskType (..), taskLocation
                    )
-import           Stack.Types.Compiler ( compilerVersionText, getGhcVersion )
+import           Stack.Types.Compiler ( getGhcVersion )
 import           Stack.Types.Config
                    ( BuildOpts (..), BuildOptsCLI (..), Config (..)
                    , EnvConfig (..), HasBuildConfig, HasConfig (..)
-                   , HasEnvConfig (..), HasPlatform (..), HasSourceMap
-                   , actualCompilerVersionL, buildOptsL, cabalVersionL
+                   , HasEnvConfig (..), HasPlatform (..), HasSourceMap, Runner
+                   , actualCompilerVersionL, buildOptsL
+                   , buildOptsMonoidBenchmarksL, buildOptsMonoidHaddockL
+                   , buildOptsMonoidInstallExesL, buildOptsMonoidTestsL
+                   , cabalVersionL, globalOptsL, globalOptsBuildOptsMonoidL
                    , installationRootDeps, installationRootLocal
                    , packageDatabaseDeps, packageDatabaseExtra
-                   , packageDatabaseLocal, stackYamlL, wantedCompilerVersionL
+                   , packageDatabaseLocal, stackYamlL
                    )
+import           Stack.Types.Config.Build
+                   ( BuildCommand (..), FileWatchOpts (..) )
 import           Stack.Types.NamedComponent ( exeComponents )
 import           Stack.Types.Package
                    ( InstallLocation (..), LocalPackage (..), Package (..)
@@ -83,26 +83,33 @@ instance Pretty CabalVersionPrettyException where
 
 instance Exception CabalVersionPrettyException
 
-data QueryException
-  = SelectorNotFound [Text]
-  | IndexOutOfRange [Text]
-  | NoNumericSelector [Text]
-  | CannotApplySelector Value [Text]
-  deriving (Show, Typeable)
-
-instance Exception QueryException where
-  displayException (SelectorNotFound sels) =
-    err "[S-4419]" "Selector not found" sels
-  displayException (IndexOutOfRange sels) =
-    err "[S-8422]" "Index out of range" sels
-  displayException (NoNumericSelector sels) =
-    err "[S-4360]" "Encountered array and needed numeric selector" sels
-  displayException (CannotApplySelector value sels) =
-    err "[S-1711]" ("Cannot apply selector to " ++ show value) sels
-
--- | Helper function for 'QueryException' instance of 'Show'
-err :: String -> String -> [Text] -> String
-err msg code sels = "Error: " ++ code ++ "\n" ++ msg ++ ": " ++ show sels
+-- | Helper for build and install commands
+buildCmd :: BuildOptsCLI -> RIO Runner ()
+buildCmd opts = do
+  when (any (("-prof" `elem`) . fromRight [] . parseArgs Escaping) (boptsCLIGhcOptions opts)) $
+    prettyThrowIO GHCProfOptionInvalid
+  local (over globalOptsL modifyGO) $
+    case boptsCLIFileWatch opts of
+      FileWatchPoll -> fileWatchPoll (inner . Just)
+      FileWatch -> fileWatch (inner . Just)
+      NoFileWatch -> inner Nothing
+ where
+  inner ::
+       Maybe (Set (Path Abs File) -> IO ())
+    -> RIO Runner ()
+  inner setLocalFiles = withConfig YesReexec $ withEnvConfig NeedTargets opts $
+      Stack.Build.build setLocalFiles
+  -- Read the build command from the CLI and enable it to run
+  modifyGO =
+    case boptsCLICommand opts of
+      Test -> set (globalOptsBuildOptsMonoidL.buildOptsMonoidTestsL) (Just True)
+      Haddock ->
+        set (globalOptsBuildOptsMonoidL.buildOptsMonoidHaddockL) (Just True)
+      Bench ->
+        set (globalOptsBuildOptsMonoidL.buildOptsMonoidBenchmarksL) (Just True)
+      Install ->
+        set (globalOptsBuildOptsMonoidL.buildOptsMonoidInstallExesL) (Just True)
+      Build -> id -- Default case is just Build
 
 -- | Build.
 --
@@ -338,70 +345,6 @@ loadPackage loc flags ghcOptions cabalConfigOpts = do
         , packageConfigPlatform = platform
         }
   resolvePackage pkgConfig <$> loadCabalFileImmutable loc
-
--- | Query information about the build and print the result to stdout in YAML format.
-queryBuildInfo :: HasEnvConfig env
-               => [Text] -- ^ selectors
-               -> RIO env ()
-queryBuildInfo selectors0 =
-      rawBuildInfo
-  >>= select id selectors0
-  >>= liftIO . TIO.putStrLn . addGlobalHintsComment . decodeUtf8 . Yaml.encode
- where
-  select _ [] value = pure value
-  select front (sel:sels) value =
-    case value of
-      Object o ->
-        case KeyMap.lookup (Key.fromText sel) o of
-          Nothing -> throwIO $ SelectorNotFound sels'
-          Just value' -> cont value'
-      Array v ->
-        case decimal sel of
-          Right (i, "")
-            | i >= 0 && i < V.length v -> cont $ v V.! i
-            | otherwise -> throwIO $ IndexOutOfRange sels'
-          _ -> throwIO $ NoNumericSelector sels'
-      _ -> throwIO $ CannotApplySelector value sels'
-   where
-    cont = select (front . (sel:)) sels
-    sels' = front [sel]
-  -- Include comments to indicate that this portion of the "stack
-  -- query" API is not necessarily stable.
-  addGlobalHintsComment
-    | null selectors0 = T.replace globalHintsLine ("\n" <> globalHintsComment <> globalHintsLine)
-    -- Append comment instead of pre-pending. The reasoning here is
-    -- that something *could* expect that the result of 'stack query
-    -- global-hints ghc-boot' is just a string literal. Seems easier
-    -- for to expect the first line of the output to be the literal.
-    | ["global-hints"] `isPrefixOf` selectors0 = (<> ("\n" <> globalHintsComment))
-    | otherwise = id
-  globalHintsLine = "\nglobal-hints:\n"
-  globalHintsComment = T.concat
-    [ "# Note: global-hints is experimental and may be renamed / removed in the future.\n"
-    , "# See https://github.com/commercialhaskell/stack/issues/3796"
-    ]
--- | Get the raw build information object
-rawBuildInfo :: HasEnvConfig env => RIO env Value
-rawBuildInfo = do
-  locals <- projectLocalPackages
-  wantedCompiler <- view $ wantedCompilerVersionL.to (utf8BuilderToText . display)
-  actualCompiler <- view $ actualCompilerVersionL.to compilerVersionText
-  pure $ object
-    [ "locals" .= Object (KeyMap.fromList $ map localToPair locals)
-    , "compiler" .= object
-        [ "wanted" .= wantedCompiler
-        , "actual" .= actualCompiler
-        ]
-    ]
- where
-  localToPair lp =
-    (Key.fromText $ T.pack $ packageNameString $ packageName p, value)
-   where
-    p = lpPackage lp
-    value = object
-      [ "version" .= CabalString (packageVersion p)
-      , "path" .= toFilePath (parent $ lpCabalFile lp)
-      ]
 
 checkComponentsBuildable :: MonadThrow m => [LocalPackage] -> m ()
 checkComponentsBuildable lps =
