+++ src/Stack/IDE.hs
@@ -1,10 +1,12 @@
 {-# LANGUAGE NoImplicitPrelude   #-}
 {-# LANGUAGE OverloadedStrings   #-}
 
--- | Functions for IDEs.
+-- | Types and functions related to Stack's @ide@ command.
 module Stack.IDE
   ( OutputStream (..)
   , ListPackagesCmd (..)
+  , idePackagesCmd
+  , ideTargetsCmd
   , listPackages
   , listTargets
   ) where
@@ -13,20 +15,42 @@ import qualified Data.Map as Map
 import qualified Data.Set as Set
 import qualified Data.Text as T
 import           Stack.Prelude
+import           Stack.Runners
+                   ( ShouldReexec (..), withBuildConfig, withConfig )
 import           Stack.Types.Config
-                   ( BuildConfig (..), HasBuildConfig (..), ppComponents )
+                   ( BuildConfig (..), HasBuildConfig (..), Runner
+                   , ppComponents
+                   )
 import           Stack.Types.NamedComponent
                    ( NamedComponent, renderPkgComponent )
 import           Stack.Types.SourceMap ( ProjectPackage (..), SMWanted (..) )
 import           System.IO ( putStrLn )
 
+-- Type representing output channel choices for the @stack ide packages@ and
+-- @stack ide targets@ commands.
 data OutputStream
   = OutputLogInfo
+    -- ^ To the same output channel as other log information.
   | OutputStdout
+    -- ^ To the standard output channel.
 
+-- Type representing output choices for the @stack ide packages@ command.
 data ListPackagesCmd
   = ListPackageNames
+    -- ^ Package names.
   | ListPackageCabalFiles
+    -- ^ Paths to Cabal files.
+
+-- | Function underlying the @stack ide packages@ command. List packages in the
+-- project.
+idePackagesCmd :: (OutputStream, ListPackagesCmd) -> RIO Runner ()
+idePackagesCmd =
+  withConfig NoReexec . withBuildConfig . uncurry listPackages
+
+-- | Function underlying the @stack ide targets@ command. List targets in the
+-- project.
+ideTargetsCmd :: OutputStream -> RIO Runner ()
+ideTargetsCmd = withConfig NoReexec . withBuildConfig . listTargets
 
 outputFunc :: HasTerm env => OutputStream -> String -> RIO env ()
 outputFunc OutputLogInfo = prettyInfo . fromString
